Renamed Welcomeand Result to TheWelcomeSite and TheResultSide and fixed dependencies.
